+++ jvector/src/com_untangle_jvector_Relay.c
@@ -29,7 +29,7 @@
 JNIEXPORT jlong JNICALL Java_com_untangle_jvector_Relay_relay_1create
     (JNIEnv * env, jobject _this)
 {
-    return (jlong)relay_create();
+    return (jlong)(uintptr_t)relay_create();
 }
 
 /*
@@ -40,7 +40,7 @@ JNIEXPORT jlong JNICALL Java_com_untangle_jvector_Relay_relay_1create
 JNIEXPORT void JNICALL Java_com_untangle_jvector_Relay_relay_1free
     (JNIEnv * env, jobject _this, jlong relay_ptr)
 {
-    relay_free((relay_t*) relay_ptr);
+    relay_free((relay_t*)(uintptr_t)relay_ptr);
 }
 
 /*
@@ -51,7 +51,7 @@ JNIEXPORT void JNICALL Java_com_untangle_jvector_Relay_relay_1free
 JNIEXPORT void JNICALL Java_com_untangle_jvector_Relay_relay_1set_1src
     (JNIEnv * env, jobject _this, jlong relay_ptr, jlong src_ptr)
 {
-    relay_set_src((relay_t*) relay_ptr, (source_t*)src_ptr);
+    relay_set_src((relay_t*)(uintptr_t)relay_ptr, (source_t*)(uintptr_t)src_ptr);
 
 }
 
@@ -63,6 +63,6 @@ JNIEXPORT void JNICALL Java_com_untangle_jvector_Relay_relay_1set_1src
 JNIEXPORT void JNICALL Java_com_untangle_jvector_Relay_relay_1set_1snk
     (JNIEnv * env, jobject _this, jlong relay_ptr, jlong snk_ptr)
 {
-    relay_set_snk((relay_t*) relay_ptr, (sink_t*)snk_ptr);
+    relay_set_snk((relay_t*)(uintptr_t)relay_ptr, (sink_t*)(uintptr_t)snk_ptr);
 }
